all migrated

+++ diary-front/src/components/TodoObject.tsx
@@ -1,23 +1,41 @@
 import './TodoObject.css';
 
+import { Avatar, Button, Checkbox, List, message, Modal } from 'antd';
 import React from 'react';
 import { connect } from 'react-redux';
-import { message, Avatar, Button, List, Checkbox, Modal } from 'antd';
 
-import api from 'utils/api';
+import { ReduxState, User } from 'reducers';
+import { dispatch } from 'reducers/store';
+import api, { DeleteTodoResponse, ErrResponse, Todo } from 'utils/api';
 import AddTodoFormContainer from './AddTodoFormContainer';
 
-class TodoObject extends React.Component {
-  state = { editVisible: false };
+class Props {
+  public todo: Todo;
+  public onCheckChange?: (e: any) => void;
+}
+class ReduxProps {
+  public user: User | null;
+}
+class State {
+  public editVisible: boolean = false;
+}
+class TodoObject extends React.Component<Props & ReduxProps, State> {
+  constructor(props: Props & ReduxProps) {
+    super(props);
+    this.state = new State();
+  }
 
-  deleteTodo() {
+  public deleteTodo() {
     const { todo, user } = this.props;
+    if (!user) {
+      return;
+    }
     api.deleteTodo({ data: { owner: user.username, todo } }).then(
-      (data) => {
+      (data: DeleteTodoResponse & ErrResponse) => {
         if (data.err) {
           message.warn('' + data.err);
         } else if (data.data.todo) {
-          this.props.dispatch({
+          dispatch({
             type: 'DELETE_TODO',
             payload: { todo: data.data.todo },
           });
@@ -29,7 +47,7 @@ class TodoObject extends React.Component {
     );
   }
 
-  render() {
+  public render() {
     const { todo, onCheckChange } = this.props;
     return (
       <List.Item
@@ -37,6 +55,7 @@ class TodoObject extends React.Component {
         actions={[
           <Button
             className="editButton"
+            key="edit"
             icon="edit"
             size="large"
             onClick={() =>
@@ -47,6 +66,7 @@ class TodoObject extends React.Component {
           />,
           <Button
             className="deleteButton"
+            key="delete"
             icon="delete"
             type="danger"
             size="large"
@@ -59,7 +79,11 @@ class TodoObject extends React.Component {
               });
             }}
           />,
-          <Checkbox defaultChecked={todo.check} onChange={onCheckChange} />,
+          <Checkbox
+            key="check"
+            defaultChecked={todo.check}
+            onChange={onCheckChange}
+          />,
         ]}
       >
         <List.Item.Meta
@@ -98,7 +122,7 @@ class TodoObject extends React.Component {
   }
 }
 
-export default connect((state) => {
+export default connect((state: ReduxState) => {
   return {
     user: state.user,
   };
